Refatora código em diferentes arquivos
